Refactored Util to put all V2 Message related utils in separate class
