+++ hnsecure/src/main/java/ca/bc/gov/hlth/hnsecure/parsing/Util.java
@@ -72,81 +72,6 @@ public final class Util {
 		return decodedString;
 	}
 
-	/**
-	 * This method is used to get the receiving application from a HL7 message.
-	 * 
-	 * @param hlMsg
-	 * @return the receiving application
-	 */
-	public static String getReceivingApp(String hlMsg) {
-
-		String recApp = "";
-
-		if (hlMsg == null || hlMsg.isEmpty()) {
-			return recApp;
-		}
-		String[] hl7Fields = hlMsg.split(DOUBLE_BACKSLASH + HL7_DELIMITER);
-		if (hl7Fields.length > 4) {
-			recApp = hl7Fields[4];
-		}
-		return recApp;
-	}
-
-	/**
-	 * returns the message type based on the HL7 message.
-	 * 
-	 * Note: the pharmaNet message with the message type ZPN
-	 * 
-	 * @param hlMsg
-	 * @return
-	 */
-	public static String getMsgType(String hlMsg) {
-
-		String msgType = "";
-
-		if (StringUtils.isEmpty(hlMsg)) {
-			return msgType;
-		}
-
-		String[] hl7MessageAtt = hlMsg.split(DOUBLE_BACKSLASH + HL7_DELIMITER);
-		if (hl7MessageAtt.length > 8) {
-			msgType = hl7MessageAtt[8];
-			// When response is generated, acknowledgment identifier is added at MSH(8)
-			if (msgType.equals(ACK) && hl7MessageAtt.length > 9) {
-				msgType = hl7MessageAtt[9];
-			}
-		}
-		// there is a special case for R50 message which the value of MSH.8 is
-		// "R50^Z05".
-		if (msgType != null && !msgType.isEmpty() && msgType.contains(R50_SPEC_CHAR)) {
-			int index = msgType.indexOf(R50_SPEC_CHAR);
-			msgType = msgType.substring(0, index);
-		}
-		return msgType;
-	}
-	
-	/**
-	 * returns the message id based on the HL7 message.
-	 * @param hlMsg
-	 * @return
-	 */
-	public static String getMsgId(String hlMsg) {
-
-		String msgId = "";
-		
-		if (StringUtils.isEmpty(hlMsg)) {
-			return msgId;
-		}
-		
-
-		String[] hl7MessageAtt = hlMsg.split(DOUBLE_BACKSLASH + HL7_DELIMITER);
-		if (hl7MessageAtt.length > 9) {
-			msgId = hl7MessageAtt[9];
-		}
-	
-		return msgId;
-	}
-	
 	/*
 	 * The FacilityId is the legacy way to track connected clients and is now set as
 	 * the ClientId of the client application In the access token this is the 'azp'
@@ -166,84 +91,6 @@ public final class Util {
 			}
 		}
 		return clientId;
-	}
-
-	/**
-	 * Checks if a segment is present in incoming HL7v2 message
-	 * 
-	 * @param v2Message
-	 * @param segmentType
-	 * @return
-	 */
-	public static boolean isSegmentPresent(String v2Message, String segmentType) {
-
-		String[] v2DataLines_Pharmanet = v2Message.split(LINE_BREAK);
-
-		for (String segment : v2DataLines_Pharmanet) {
-
-			if (segment.startsWith(segmentType)) {
-				String[] messageSegments = segment.split(DOUBLE_BACKSLASH + HL7_DELIMITER);
-				if (messageSegments[0].equalsIgnoreCase(segmentType)) {
-					return true;
-				}
-			}
-
-		}
-		return false;
-	}
-	
-	/**
-	 * @param v2Message
-	 * @param segmentType
-	 * @return
-	 */
-	public static String getZCBSegment(String v2Message, String segmentType) {
-		String[] v2DataLinesPharmanet = v2Message.split(LINE_BREAK);
-
-		for (String segment : v2DataLinesPharmanet) {
-
-			if (segment.startsWith(segmentType)) {
-				String[] messageSegments = segment.split(DOUBLE_BACKSLASH + HL7_DELIMITER);
-				if (messageSegments[0].equalsIgnoreCase(segmentType)) {
-					return segment;
-				}
-			}
-
-		}	
-		
-		return null;
-	}
-	
-	
-	/**
-	 * @param zcbSegment
-	 * ZCB|PharmacyId|DateTime|TraceNumber
-	 * @return Pharmacy id
-	 */
-	public static String getPharmacyId(String zcbSegment) {		
-		if(StringUtils.isNotBlank(zcbSegment)) {
-			String[] zcbDataSegment = zcbSegment.split(DOUBLE_BACKSLASH + HL7_DELIMITER);
-			if(zcbDataSegment.length >1) {
-				return zcbDataSegment[1];
-			}
-		}
-		return "";
-	}
-	
-	
-	/**
-	 * @param zcbSegment
-	 * ZCB|PharmacyId|DateTime|TraceNumber
-	 * @return trace number
-	 */
-	public static String getTraceNumber(String zcbSegment) {		
-		if(StringUtils.isNotEmpty(zcbSegment)) {
-			String[] zcbDataSegment = zcbSegment.split(DOUBLE_BACKSLASH + HL7_DELIMITER);
-			if(zcbDataSegment.length >3) {
-				return zcbDataSegment[3];
-			}
-		}
-		return "";
 	}
 	
 	/**
